scripts/export_requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
optimum>=1.17.0
onnx>=1.15.0
onnxruntime>=1.17.0
pillow>=10.0.0
numpy<2.0.0